+++ apps/gamrywb_CSC_app.m
@@ -525,11 +525,7 @@ function varargout = gamrywb_CSC_app(varargin)
     end
 
     function addLog(msg)
-        timestamp = datestr(now,'HH:MM:SS');
-        old = txtLog.Value;
-        old{end+1} = sprintf('[%s] %s', timestamp, char(msg));
-        txtLog.Value = old;
-        drawnow limitrate
+        gamrywb.ui.appendLog(txtLog, msg);
     end
 end