+++ apps/gamrywb_EIS_app.m
@@ -397,11 +397,7 @@ function varargout = gamrywb_EIS_app(varargin)
     end
 
     function addLog(msg)
-        ts = datestr(now, 'HH:MM:SS');
-        old = txtLog.Value;
-        old{end+1} = sprintf('[%s] %s', ts, char(msg));
-        txtLog.Value = old;
-        drawnow limitrate
+        gamrywb.ui.appendLog(txtLog, msg);
     end
 end