+++ apps/gamrywb_VTResistance_app.m
@@ -610,11 +610,7 @@ function varargout = gamrywb_VTResistance_app(varargin)
     end
 
     function addLog(msg)
-        ts = datestr(now,'HH:MM:SS');
-        old = txtLog.Value;
-        old{end+1} = sprintf('[%s] %s', ts, char(msg));
-        txtLog.Value = old;
-        drawnow limitrate
+        gamrywb.ui.appendLog(txtLog, msg);
     end
 
 end